Python 3: replace long with int
* Replace 123L with 123: "123L" is a syntax error on Python 3
* Replace long() with int()
* Replace (int, long) with six.integer_types
* Make long type conditional in META_TYPE_MAP
  (ceilometer/storage/sqlalchemy/utils.py)

+++ ceilometer/compute/virt/hyperv/inspector.py
@@ -41,7 +41,7 @@ class HyperVInspector(virt_inspector.Inspector):
         cpu_percent_used = (cpu_clock_used /
                             float(host_cpu_clock * cpu_count))
         # Nanoseconds
-        cpu_time = (long(uptime * cpu_percent_used) * units.k)
+        cpu_time = (int(uptime * cpu_percent_used) * units.k)
 
         return virt_inspector.CPUStats(number=cpu_count, time=cpu_time)
